ktime fixes
[ 2214.041273] ================================================================================
[ 2214.042111] UBSan: Undefined behaviour in ../kernel/time/hrtimer.c:293:16
[ 2214.042721] signed integer overflow:
[ 2214.043059] 2214041269968 + 9223372036854775807 cannot be represented in type 'long long int'
[ 2214.043871] CPU: 1 PID: 6147 Comm: trinity-c1 Not tainted 3.17.0-rc7+ #39
[ 2214.044477] Hardware name: QEMU Standard PC (i440FX + PIIX, 1996), BIOS rel-1.7.5-0-ge51488c-20140602_164612-nilsson.home.kraxel.org 04/01/2014
[ 2214.045862]  0000000000000000 0000000000000000 000000000000002b ffff8800b895fe30
[ 2214.046714]  ffffffff82a30940 7fffffffffffffff ffff8800b895fe40 ffffffff819502e9
[ 2214.047552]  ffff8800b895fec0 ffffffff81950b0e 3231343034313232 ffff003836393936
[ 2214.048488] Call Trace:
[ 2214.048777] dump_stack (/home/andrew/linux/ubsan_x86/.lib/dump_stack.c:52)
[ 2214.049337] ubsan_epilogue (/home/andrew/linux/ubsan_x86/.lib/ubsan.c:122)
[ 2214.049904] handle_overflow (/home/andrew/linux/ubsan_x86/.lib/ubsan.c:153 (discriminator 4))
[ 2214.050508] __ubsan_handle_add_overflow (/home/andrew/linux/ubsan_x86/.lib/ubsan.c:161)
[ 2214.051219] ktime_add_safe (/home/andrew/linux/ubsan_x86/.kernel/time/hrtimer.c:293)
[ 2214.051800] SyS_futex (/home/andrew/linux/ubsan_x86/.kernel/futex.c:2983 /home/andrew/linux/ubsan_x86/.kernel/futex.c:2964)
[ 2214.052369] ? alarm_setitimer (/home/andrew/linux/ubsan_x86/.kernel/time/itimer.c:271)
[ 2214.052976] system_call_fastpath (/home/andrew/linux/ubsan_x86/.arch/x86/kernel/entry_64.S:423)
[ 2214.053665] ================================================================================

+++ include/linux/ktime.h
@@ -57,11 +57,11 @@ static inline ktime_t ktime_set(const s64 secs, const unsigned long nsecs)
 
 /* Subtract two ktime_t variables. rem = lhs -rhs: */
 #define ktime_sub(lhs, rhs) \
-		({ (ktime_t){ .tv64 = (lhs).tv64 - (rhs).tv64 }; })
+		({ (ktime_t){ .tv64 = (u64)(lhs).tv64 - (u64)(rhs).tv64 }; })
 
 /* Add two ktime_t variables. res = lhs + rhs: */
 #define ktime_add(lhs, rhs) \
-		({ (ktime_t){ .tv64 = (lhs).tv64 + (rhs).tv64 }; })
+		({ (ktime_t){ .tv64 = (u64)(lhs).tv64 + (u64)(rhs).tv64 }; })
 
 /*
  * Add a ktime_t variable and a scalar nanosecond value.
